+++ packages/node/test/node/ClientNodeTest.ts
@@ -192,14 +192,14 @@ describe("ClientNode", () => {
 
         // *** INVOCATION ***
 
-        await ep1.commandsOf(OnOffClient).toggle();
+        await MockTime.resolve(ep1.commandsOf(OnOffClient).toggle());
 
         // *** UPDATE ***
 
         await MockTime.resolve(receivedUpdate);
 
         // *** Test another command also in the feature-set ***
-        await ep1.commandsOf(OnOffClient).offWithEffect({ effectIdentifier: 0, effectVariant: 0 });
+        await MockTime.resolve(ep1.commandsOf(OnOffClient).offWithEffect({ effectIdentifier: 0, effectVariant: 0 }));
     });
 
     it("decommissions", async () => {
@@ -345,7 +345,7 @@ describe("ClientNode", () => {
             ep1.eventsOf(OnOffClient).onOff$Changed.once(resolve);
         });
 
-        await ep1.commandsOf(OnOffClient).toggle();
+        await MockTime.resolve(ep1.commandsOf(OnOffClient).toggle());
 
         await MockTime.resolve(toggled);